@@ -17,20 +17,20 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^7.4",
+        "php": "^8.1",
         "ext-json": "*",
-        "spatie/laravel-mailcoach": "^2.0",
-        "spatie/laravel-medialibrary": "^8.0"
+        "illuminate/support": "^9.0|^10.0",
+        "spatie/laravel-mailcoach": "^6.0"
     },
     "require-dev": {
-        "orchestra/testbench": "^5.0",
-        "phpunit/phpunit": "^9.0",
-        "symfony/var-dumper": "^5.0"
+        "orchestra/testbench": "^7.0|^8.0",
+        "phpunit/phpunit": "^9.5",
+        "symfony/var-dumper": "^6.0"
     },
     "repositories": [
         {
             "type": "composer",
-            "url": "https://satis.mailcoach.app"
+            "url": "https://satis.spatie.be"
         }
     ],
     "autoload": {
@@ -40,7 +40,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Spatie\\MailcoachUnlayer\\Tests\\": "tests"
+            "Spatie\\MailcoachUnlayer\\Tests\\": "tests",
+            "Spatie\\Mailcoach\\Database\\Factories\\": "vendor/spatie/laravel-mailcoach/database/factories"
         }
     },
     "scripts": {
