@@ -2,33 +2,25 @@
 
 namespace Spatie\MailcoachUnlayer;
 
-use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
+use Livewire\Livewire;
+use Spatie\Mailcoach\Mailcoach;
 
 class MailcoachUnlayerServiceProvider extends ServiceProvider
 {
     public function boot()
     {
-        $this->loadViewsFrom(__DIR__.'/../resources/views', 'mailcoach-unlayer');
+        $this->loadViewsFrom(__DIR__ . '/../resources/views', 'mailcoach-unlayer');
+        $this->loadJsonTranslationsFrom(__DIR__ . '/../resources/lang');
 
         if ($this->app->runningInConsole()) {
             $this->publishes([
-                __DIR__.'/../resources/views' => base_path('resources/views/vendor/mailcoach/unlayer'),
+                __DIR__ . '/../resources/views' => base_path('resources/views/vendor/mailcoach-unlayer'),
             ], 'mailcoach-unlayer-views');
-
-            if (!class_exists('CreateMailcoachUnlayerTables')) {
-                $this->publishes([
-                    __DIR__ . '/../database/migrations/create_mailcoach_unlayer_tables.php.stub' => database_path('migrations/' . date('Y_m_d_His', time()) . '_create_mailcoach_unlayer_tables.php'),
-                ], 'mailcoach-unlayer-migrations');
-            }
         }
 
-        Route::macro('mailcoachUnlayer', function (string $url = '') {
-            Route::prefix($url)->group(function () {
-                $middlewareClasses = config('mailcoach.middleware');
+        Livewire::component('mailcoach-unlayer::editor', UnlayerEditor::class);
 
-                Route::middleware($middlewareClasses)->prefix('')->group(__DIR__ . '/../routes/api.php');
-            });
-        });
+        Mailcoach::editorScript(UnlayerEditor::class, 'https://editor.unlayer.com/embed.js');
     }
 }
