@@ -2,10 +2,9 @@
 
 namespace Spatie\MailcoachUnlayer\Tests;
 
-use CreateMailcoachTables;
-use CreateMailcoachUnlayerTables;
-use CreateMediaTable;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Facades\Route;
+use Livewire\LivewireServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 use Spatie\Mailcoach\MailcoachServiceProvider;
 use Spatie\MailcoachUnlayer\MailcoachUnlayerServiceProvider;
@@ -17,10 +16,11 @@ abstract class TestCase extends Orchestra
     {
         parent::setUp();
 
-        $this->withFactories(__DIR__.'/../database/factories');
-        $this->withFactories(__DIR__.'/../vendor/spatie/laravel-mailcoach/database/factories');
+        Factory::guessFactoryNamesUsing(
+            fn (string $modelName) => 'Spatie\\Mailcoach\\Database\\Factories\\'.class_basename($modelName).'Factory'
+        );
 
-        Route::mailcoachUnlayer('mailcoachUnlayer');
+        Route::mailcoach('mailcoach');
 
         $this->withoutExceptionHandling();
     }
@@ -31,6 +31,7 @@ abstract class TestCase extends Orchestra
             MediaLibraryServiceProvider::class,
             MailcoachServiceProvider::class,
             MailcoachUnlayerServiceProvider::class,
+            LivewireServiceProvider::class,
         ];
     }
 
@@ -43,13 +44,10 @@ abstract class TestCase extends Orchestra
             'prefix' => '',
         ]);
 
-        include_once __DIR__.'/../vendor/spatie/laravel-mailcoach/database/migrations/create_mailcoach_tables.php.stub';
-        (new CreateMailcoachTables())->up();
+        $createMailcoachTables = require __DIR__.'/../../../vendor/spatie/laravel-mailcoach/database/migrations/create_mailcoach_tables.php';
+        $createMailcoachTables->up();
 
-        include_once __DIR__.'/../database/migrations/create_mailcoach_unlayer_tables.php.stub';
-        (new CreateMailcoachUnlayerTables())->up();
-
-        include_once __DIR__.'/../vendor/spatie/laravel-medialibrary/database/migrations/create_media_table.php.stub';
-        (new CreateMediaTable())->up();
+        $createMediaTable = require __DIR__.'/../../../vendor/spatie/laravel-mailcoach/database/migrations/create_media_table.php';
+        $createMediaTable->up();
     }
 }
